fix: quiz summary page ui fixes
